requirements.txt
python-telegram-bot==20.7
requests==2.31.0
humanize==4.9.0
python-dotenv==1.0.0
